+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/messages/list/MessageListController.kt
@@ -16,6 +16,7 @@
 
 package io.getstream.chat.android.ui.common.feature.messages.list
 
+import androidx.annotation.VisibleForTesting
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.channel.state.ChannelState
 import io.getstream.chat.android.client.errors.extractCause
@@ -73,6 +74,7 @@ import io.getstream.chat.android.ui.common.state.messages.list.DateSeparatorItem
 import io.getstream.chat.android.ui.common.state.messages.list.DeletedMessageVisibility
 import io.getstream.chat.android.ui.common.state.messages.list.EmptyThreadPlaceholderItemState
 import io.getstream.chat.android.ui.common.state.messages.list.GiphyAction
+import io.getstream.chat.android.ui.common.state.messages.list.HasMessageListItemState
 import io.getstream.chat.android.ui.common.state.messages.list.MessageFocusRemoved
 import io.getstream.chat.android.ui.common.state.messages.list.MessageFocused
 import io.getstream.chat.android.ui.common.state.messages.list.MessageFooterVisibility
@@ -381,10 +383,19 @@ public class MessageListController(
     private val debouncer = Debouncer(debounceMs = 200L, scope = scope)
 
     @Volatile
-    private var lastSeenMessageId: String? = null
+    private var lastSeenChannelMessageId: String? = null
 
     @Volatile
     private var lastSeenThreadMessageId: String? = null
+
+    @VisibleForTesting
+    internal var lastSeenMessageId: String?
+        get() = if (isInThread) lastSeenThreadMessageId else lastSeenChannelMessageId
+        set(value) = if (isInThread) {
+            lastSeenThreadMessageId = value
+        } else {
+            lastSeenChannelMessageId = value
+        }
 
     /**
      * We start observing messages and if the message list screen was started after searching for a message, it will
@@ -1433,21 +1444,19 @@ public class MessageListController(
      * Marks that the last message in the list as read. This also sets the unread count to 0.
      */
     private fun markLastMessageReadInternal() {
-        val itemState = messagesState.messageItems.lastOrNull { it is MessageItemState } as? MessageItemState
+        val itemState = messagesState.messageItems.lastOrNull { messageItem ->
+            messageItem is HasMessageListItemState
+        } as? HasMessageListItemState
         val messageId = itemState?.message?.id
         val messageText = itemState?.message?.text
         logger.d { "[markLastMessageRead] cid: $cid, msgId($isInThread): $messageId, msgText: \"$messageText\"" }
 
-        val lastSeenMessageId = if (isInThread) lastSeenThreadMessageId else lastSeenMessageId
+        val lastSeenMessageId = this.lastSeenMessageId
         if (lastSeenMessageId == messageId) {
             logger.w { "[markLastMessageRead] cid: $cid; rejected[$isInThread] (already seen msgId): $messageId" }
             return
         }
-        if (isInThread) {
-            this.lastSeenThreadMessageId = messageId
-        } else {
-            this.lastSeenMessageId = messageId
-        }
+        this.lastSeenMessageId = messageId
 
         cid.cidToTypeAndId().let { (channelType, channelId) ->
             if (isInThread) {